Split TagSet into abstract class in core and implementation in core_impl.
