+++ core/src/main/java/com/google/instrumentation/tags/TagSet.java
@@ -13,35 +13,13 @@
 
 package com.google.instrumentation.tags;
 
-import com.google.common.base.Preconditions;
 import com.google.instrumentation.internal.StringUtil;
-import com.google.instrumentation.internal.logging.Logger;
-import com.google.instrumentation.tags.TagKey.TagType;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.logging.Level;
 import javax.annotation.Nullable;
-import javax.annotation.concurrent.NotThreadSafe;
 
 /** A set of tags. */
-// The class is immutable, except for the logger.
-public final class TagSet {
+public abstract class TagSet {
   /** The maximum length for a string tag value. */
   public static final int MAX_STRING_LENGTH = StringUtil.MAX_LENGTH;
-
-  private final Logger logger;
-
-  // The types of the TagKey and value must match for each entry.
-  private final Map<TagKey<?>, Object> tags;
-
-  TagSet(Logger logger, Map<TagKey<?>, Object> tags) {
-    this.logger = logger;
-    this.tags = tags;
-  }
-
-  Map<TagKey<?>, Object> getTags() {
-    return tags;
-  }
 
   /**
    * Determines whether a key is present.
@@ -49,9 +27,7 @@ public final class TagSet {
    * @param key the key to look up.
    * @return {@code true} if the key is present.
    */
-  public boolean tagKeyExists(TagKey<?> key) {
-    return tags.containsKey(key);
-  }
+  public abstract boolean tagKeyExists(TagKey<?> key);
 
   // TODO(sebright): Which is better, the generic "get" method or three specialized "get" methods?
   // Should these methods take defaults or throw exceptions to handle missing keys?
@@ -63,9 +39,7 @@ public final class TagSet {
    * @return the tag value, or {@code null} if the key isn't present.
    */
   @Nullable
-  public String getStringTagValue(TagKey<String> key) {
-    return (String) tags.get(key);
-  }
+  public abstract String getStringTagValue(TagKey<String> key);
 
   /**
    * Gets a value of type {@code long}.
@@ -74,10 +48,7 @@ public final class TagSet {
    * @param defaultValue the value to return if the key is not preset.
    * @return the tag value, or the default value if the key is not present.
    */
-  public long getIntTagValue(TagKey<Long> key, long defaultValue) {
-    Long value = (Long) tags.get(key);
-    return value == null ? defaultValue : value;
-  }
+  public abstract long getIntTagValue(TagKey<Long> key, long defaultValue);
 
   /**
    * Gets a value of type {@code boolean}.
@@ -86,10 +57,7 @@ public final class TagSet {
    * @param defaultValue the value to return if the key is not preset.
    * @return the tag value, or the default value if the key is not present.
    */
-  public boolean getBooleanTagValue(TagKey<Boolean> key, boolean defaultValue) {
-    Boolean value = (Boolean) tags.get(key);
-    return value == null ? defaultValue : value;
-  }
+  public abstract boolean getBooleanTagValue(TagKey<Boolean> key, boolean defaultValue);
 
   /**
    * Gets a value of any type.
@@ -98,32 +66,17 @@ public final class TagSet {
    * @return the tag value, or {@code null} if the key isn't present.
    */
   @Nullable
-  public <T> T getTagValue(TagKey<T> key) {
-    // An unchecked cast is okay, because we validate the values when they are inserted.
-    @SuppressWarnings("unchecked")
-    T value = (T) tags.get(key);
-    return value;
-  }
+  public abstract <T> T getTagValue(TagKey<T> key);
 
-  public Builder toBuilder() {
-    return new Builder(logger, getTags());
-  }
+  /**
+   * Returns a builder based on this {@code TagSet}.
+   *
+   * @return a builder based on this {@code TagSet}.
+   */
+  public abstract Builder toBuilder();
 
   /** Builder for the {@link TagSet} class. */
-  @NotThreadSafe
-  public static final class Builder {
-    private final Logger logger;
-    private final Map<TagKey<?>, Object> tags;
-
-    private Builder(Logger logger, Map<TagKey<?>, Object> tags) {
-      this.logger = logger;
-      this.tags = new HashMap<TagKey<?>, Object>(tags);
-    }
-
-    Builder(Logger logger) {
-      this.logger = logger;
-      this.tags = new HashMap<TagKey<?>, Object>();
-    }
+  public abstract static class Builder {
 
     /**
      * Adds the key/value pair if the key is not present. If the key is present, it logs an error.
@@ -132,10 +85,7 @@ public final class TagSet {
      * @param value the value to insert for the given key.
      * @return this
      */
-    public Builder insert(TagKey<String> key, String value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_STRING);
-      return insertInternal(key, StringUtil.sanitize(value));
-    }
+    public abstract Builder insert(TagKey<String> key, String value);
 
     /**
      * Adds the key/value pair if the key is not present. If the key is present, it logs an error.
@@ -144,10 +94,7 @@ public final class TagSet {
      * @param value the value to insert for the given key.
      * @return this
      */
-    public Builder insert(TagKey<Long> key, long value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_INT);
-      return insertInternal(key, value);
-    }
+    public abstract Builder insert(TagKey<Long> key, long value);
 
     /**
      * Adds the key/value pair if the key is not present. If the key is present, it logs an error.
@@ -156,19 +103,7 @@ public final class TagSet {
      * @param value the value to insert for the given key.
      * @return this
      */
-    public Builder insert(TagKey<Boolean> key, boolean value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_BOOL);
-      return insertInternal(key, value);
-    }
-
-    private <TagValueT> Builder insertInternal(TagKey<TagValueT> key, TagValueT value) {
-      if (!tags.containsKey(key)) {
-        tags.put(key, value);
-      } else {
-        logger.log(Level.WARNING, "Tag key already exists: " + key);
-      }
-      return this;
-    }
+    public abstract Builder insert(TagKey<Boolean> key, boolean value);
 
     /**
      * Adds the key/value pair regardless of whether the key is present.
@@ -177,10 +112,7 @@ public final class TagSet {
      * @param value the value to set for the given key.
      * @return this
      */
-    public Builder set(TagKey<String> key, String value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_STRING);
-      return setInternal(key, StringUtil.sanitize(value));
-    }
+    public abstract Builder set(TagKey<String> key, String value);
 
     /**
      * Adds the key/value pair regardless of whether the key is present.
@@ -189,10 +121,7 @@ public final class TagSet {
      * @param value the value to set for the given key.
      * @return this
      */
-    public Builder set(TagKey<Long> key, long value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_INT);
-      return setInternal(key, value);
-    }
+    public abstract Builder set(TagKey<Long> key, long value);
 
     /**
      * Adds the key/value pair regardless of whether the key is present.
@@ -201,15 +130,7 @@ public final class TagSet {
      * @param value the value to set for the given key.
      * @return this
      */
-    public Builder set(TagKey<Boolean> key, boolean value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_BOOL);
-      return setInternal(key, value);
-    }
-
-    private <TagValueT> Builder setInternal(TagKey<TagValueT> key, TagValueT value) {
-      tags.put(key, value);
-      return this;
-    }
+    public abstract Builder set(TagKey<Boolean> key, boolean value);
 
     /**
      * Adds the key/value pair only if the key is already present.
@@ -218,10 +139,7 @@ public final class TagSet {
      * @param value the value to update for the given key.
      * @return this
      */
-    public Builder update(TagKey<String> key, String value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_STRING);
-      return updateInternal(key, StringUtil.sanitize(value));
-    }
+    public abstract Builder update(TagKey<String> key, String value);
 
     /**
      * Adds the key/value pair only if the key is already present.
@@ -230,10 +148,7 @@ public final class TagSet {
      * @param value the value to update for the given key.
      * @return this
      */
-    public Builder update(TagKey<Long> key, long value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_INT);
-      return updateInternal(key, value);
-    }
+    public abstract Builder update(TagKey<Long> key, long value);
 
     /**
      * Adds the key/value pair only if the key is already present.
@@ -242,17 +157,7 @@ public final class TagSet {
      * @param value the value to update for the given key.
      * @return this
      */
-    public Builder update(TagKey<Boolean> key, boolean value) {
-      Preconditions.checkArgument(key.getTagType() == TagType.TAG_BOOL);
-      return updateInternal(key, value);
-    }
-
-    private <TagValueT> Builder updateInternal(TagKey<TagValueT> key, TagValueT value) {
-      if (tags.containsKey(key)) {
-        tags.put(key, value);
-      }
-      return this;
-    }
+    public abstract Builder update(TagKey<Boolean> key, boolean value);
 
     /**
      * Removes the key if it exists.
@@ -260,13 +165,13 @@ public final class TagSet {
      * @param key the key to look up.
      * @return this
      */
-    public Builder clear(TagKey<?> key) {
-      tags.remove(key);
-      return this;
-    }
+    public abstract Builder clear(TagKey<?> key);
 
-    public TagSet build() {
-      return new TagSet(logger, new HashMap<TagKey<?>, Object>(tags));
-    }
+    /**
+     * Creates a {@code TagSet} from this builder.
+     *
+     * @return a {@code TagSet} with the same tags as this builder.
+     */
+    public abstract TagSet build();
   }
 }
