+++ core/src/main/java/com/google/instrumentation/tags/TagSetFactory.java
@@ -13,47 +13,22 @@
 
 package com.google.instrumentation.tags;
 
-import com.google.instrumentation.internal.logging.Logger;
-import com.google.instrumentation.internal.logging.LoggerFactory;
-
 /**
  * Factory for {@link TagSet}s.
  */
-public final class TagSetFactory {
-
-  private final TagSet emptyTagSet;
-  private final Logger tagSetLogger;
-
-  /**
-   * Creates a {@code TagSetFactory} which creates {@code TagSet}s that log using the given
-   * {@code LoggerFactory}.
-   *
-   * @param loggerFactory a {@code LoggerFactory} used for {@code TagSet}-related logging.
-   */
-  static TagSetFactory create(LoggerFactory loggerFactory) {
-    return new TagSetFactory(loggerFactory);
-  }
-
-  private TagSetFactory(LoggerFactory loggerFactory) {
-    this.tagSetLogger = loggerFactory.getLogger(TagSet.class.getName());
-    this.emptyTagSet = new TagSet.Builder(tagSetLogger).build();
-  }
+public abstract class TagSetFactory {
 
   /**
    * Returns an empty builder.
    *
    * @return an empty builder.
    */
-  public TagSet.Builder builder() {
-    return new TagSet.Builder(tagSetLogger);
-  }
+  public abstract TagSet.Builder builder();
 
   /**
    * Returns an empty {@code TagSet}.
    *
    * @return an empty {@code TagSet}.
    */
-  public TagSet empty() {
-    return emptyTagSet;
-  }
+  public abstract TagSet empty();
 }
