+++ core_impl/src/main/java/com/google/instrumentation/internal/logging/Logger.java
