+++ core_impl/src/main/java/com/google/instrumentation/stats/StatsManagerImplBase.java
@@ -22,8 +22,7 @@ import com.google.instrumentation.stats.MutableView.MutableDistributionView;
 import com.google.instrumentation.stats.MutableView.MutableIntervalView;
 import com.google.instrumentation.stats.ViewDescriptor.DistributionViewDescriptor;
 import com.google.instrumentation.stats.ViewDescriptor.IntervalViewDescriptor;
-import com.google.instrumentation.tags.TagSetFactory;
-import com.google.instrumentation.tags.TagSetFactoryAccessor;
+import com.google.instrumentation.tags.TagSetFactoryImpl;
 
 /**
  * Native Implementation of {@link StatsManager}.
@@ -41,7 +40,7 @@ class StatsManagerImplBase extends StatsManager {
   // created in the constructor.  Multiple initializations are okay.
   private volatile StatsContextFactoryImpl statsContextFactory;
 
-  private final TagSetFactory tagSetFactory;
+  private final TagSetFactoryImpl tagSetFactory;
 
   StatsManagerImplBase(EventQueue queue, Clock clock) {
     this(queue, clock, new JavaLoggerFactory());
@@ -50,7 +49,7 @@ class StatsManagerImplBase extends StatsManager {
   StatsManagerImplBase(EventQueue queue, Clock clock, LoggerFactory loggerFactory) {
     this.queue = queue;
     this.clock = clock;
-    this.tagSetFactory = TagSetFactoryAccessor.createTagSetFactory(loggerFactory);
+    this.tagSetFactory = TagSetFactoryImpl.create(loggerFactory);
   }
 
   @Override
@@ -98,7 +97,7 @@ class StatsManagerImplBase extends StatsManager {
   }
 
   @Override
-  TagSetFactory getTagSetFactory() {
+  TagSetFactoryImpl getTagSetFactory() {
     return tagSetFactory;
   }
