--- core_impl/src/main/java/com/google/instrumentation/tags/TagSetFactoryAccessor.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2017, Google Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.instrumentation.tags;
-
-import com.google.instrumentation.internal.logging.LoggerFactory;
-import com.google.instrumentation.tags.TagSetFactory;
-
-/**
- * Class for exposing {@link TagSetFactory}'s factory method to the stats package.
- */
-public class TagSetFactoryAccessor {
-  public static TagSetFactory createTagSetFactory(LoggerFactory loggerFactory) {
-    return TagSetFactory.create(loggerFactory);
-  }
-}
